+++ packages/react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @valbuild/react
 
+## 0.136.4
+
+### Patch Changes
+
+- Updated dependencies [[`d19b1f6`](https://github.com/valbuild/val/commit/d19b1f67360b057e63536ac020f7109bd2496389)]:
+  - @valbuild/ui@0.136.4
+
+## 0.136.3
+
+### Patch Changes
+
+- Updated dependencies [[`38d24f7`](https://github.com/valbuild/val/commit/38d24f723639ee109d1007854507c4571bc82575), [`3b19534`](https://github.com/valbuild/val/commit/3b19534bb7aa16452a104906d9594083c3bc3e89), [`fa28164`](https://github.com/valbuild/val/commit/fa28164dfd1e05ff53eccf375334741d857fd2ef)]:
+  - @valbuild/ui@0.136.3
+  - @valbuild/shared@0.136.3
+
 ## 0.136.2
 
 ### Patch Changes